Canvas: add SVG compatibility checks
Adds a function that returns a informational warning if the canvas contains any boxes and/or effects not compatible with SVG export. Can also include notes or recommendations.

This will be used by the SVG exporter to give the user feedback on the project compatibility before exporting.

It's not perfect, but should give the user enough information to adjust the project for export.

Ref: #174