HiDPI warning dialog
Warn with dialog if requested icon size is not available, this should be considered a critical issue.

Note that on Windows if you change display scaling without restarting it, icon sizes will be wacky in Friction, this is an issue with Windows and/or Qt, probably a combination.

So, if you get this warning on Windows you should probably restart.