UI fix: make sure default tool mode is set
Make sure default tool mode (object) is set in the UI before starting a new project, else you might end up with the wrong tool mode selected.

Example: you are in text mode, you (for whatever reason) make or load a new project, the text tool is still "active" (as in selected in the UI) but you are actually in object mode since that is the default mode for any new project.