macOS cache hack
Yeah... I hate hacks, but here we are. I have no solution for the cache issues on macOS, so it's either this hack or no macOS support.

All cache will now be cleared on document changed, regardless of frame state. This is sub optimal, as we only want to render changed frames, but from a users perspective it's better to clear all than none.