flatpak: workaround limited portal support
If running as flatpak limit save image sequence to XDG Pictures/Documents.

There are probably other places where this also should be checked for, can we import sequences?

There is no way for us to support 'org.freedesktop.portal.FileChooser::SaveFiles' as-is.

- We allow 'xdg-documents' for SVG preview, plugins and image sequences
- We allow 'xdg-pictures' for image sequences

The rest of Friction can use portal to load and save to any directory.